Changes to get all users to signin and added admin roles

+++ src/app/api/blog/delete/[id]/route.ts
@@ -9,29 +9,30 @@ export async function DELETE(
 ) {
   try {
     const session = await getServerSession(authOptions);
-
     if (!session?.user?.email) {
       return NextResponse.json(
         { message: "Unauthorized" },
         { status: 401 }
       );
     }
-
+    if ((session.user as any).role !== 'admin') {
+      return NextResponse.json(
+        { message: "Forbidden: Admins only" },
+        { status: 403 }
+      );
+    }
     // Access the `id` from the `params` object
     const { id } = params;
-
     // Delete the post with the specified ID
     await prisma.post.delete({
       where: { id },
     });
-
     return NextResponse.json({
       success: true,
       message: "Post deleted successfully",
     });
   } catch (error) {
     console.error("Error deleting post:", error);
-
     return NextResponse.json(
       { message: "Failed to delete post" },
       { status: 500 }
